WAGE: Move all console-related stuff to gui-console.cpp
